+++ src/Symfony/Component/Console/Command/CompleteCommand.php
@@ -34,16 +34,6 @@ final class CompleteCommand extends Command
 {
     public const COMPLETION_API_VERSION = '1';
 
-    /**
-     * @deprecated since Symfony 6.1
-     */
-    protected static $defaultName = '|_complete';
-
-    /**
-     * @deprecated since Symfony 6.1
-     */
-    protected static $defaultDescription = 'Internal command to provide shell completion suggestions';
-
     private $completionOutputs;
 
     private $isDebug = false;