+++ src/Symfony/Component/Dotenv/Command/DebugCommand.php
@@ -29,16 +29,6 @@ use Symfony\Component\Dotenv\Dotenv;
 #[AsCommand(name: 'debug:dotenv', description: 'Lists all dotenv files with variables and values')]
 final class DebugCommand extends Command
 {
-    /**
-     * @deprecated since Symfony 6.1
-     */
-    protected static $defaultName = 'debug:dotenv';
-
-    /**
-     * @deprecated since Symfony 6.1
-     */
-    protected static $defaultDescription = 'Lists all dotenv files with variables and values';
-
     private $kernelEnvironment;
     private $projectDirectory;